Add new person entityData when necessary
